ct: Add L0_read_path module
The module contains placeholder_extent and placeholder_extent reader
and uses them to implement a 'l0_fetch_handler' component. The
'l0_fetch_handler' is a pipeline component that monitors the read
pipeline and resolves L0 fetch requests which target the right offset
range. The 'resolver' is just creating the 'placeholder_extent_reader'
under the hood.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>