dt/crash_tracker: test crash reports with signals
This adds a test to verify that crash reports captured while handling
crash signals (segfaults, sigabrt, sigill) are correctly being reported.
It also verifies that the generated crash reports can be read back later
and are correctly printed when the crash loop limit is reached.

This requires exposing `_tolerate_crashes` from `RedpandaService` to
prevent test failures on crash-detecting checks, since for this test,
the crashes are expected.