iceberg: optionally tag appended snapshot
Adds the capability to tag a snapshot when appended. This will be useful
for pinning a snapshot to the table, since tagged snapshots don't expire
by default. We'd like to keep some metadata around in Redpanda-generated
snapshots for exactly once delivery, and this helps ensure that metadata
don't get wiped out.