bazel: specify -march compile flags
There isn't a great way to auto add --copt flags based on architecture,
so I am overriding the flags in toolchains_llvm to specify the correct
-march, as those configurations are already platform specific.

The default flags are preserved, I've only added the -march flags here
(since this attribute overrides the defaults).
We may want to migrate some flags we specify in .bazelrc to here and
consolidate them, but for now this will do. I don't think anyone is
still using their system toolchain for bazel, but these flags aren't
applied there.

The diff of the `compile_commands.json` file shows that only `-march`
was added to the list of compilation flags.