parquet: move page flushing up to parquet writer
hunter detected a performance regression in the datalake benchmark due
the change to flush pages. My investigation made me realize this caused
an extra move for iobufs (which are the string values inside all of
these containers), which is the cause of the performance regression.

The move is caused by a coroutine frame addition in column_writer::add,
and we can remove it by not coroutizing that function and pushing the
flush_page call into the caller of column_writer::add, which is already
a coroutine. This drops the benchmarks back to about to the same level
as before.