`storage`: add `dirty_and_closed_bytes_bookkeeping` test
This test uses a fixed number of operations that influence the number of
dirty and closed bytes in the log, chosen at random.

The book-kept `_closed_segment_bytes` and `_dirty_segment_bytes` are checked
after every operation in order to ensure we are accurately book keeping
these values in `disk_log_impl`.