kafka: make produce_consume fixture friend of ...
... kafka_probe. Instead of using public accessor functions for
internals, make the fixture that needs this access a friend. This limits
the exposure of internal to that fixture only (though of course that
fixture now has more access, if it was feeling naughty) and avoids
adding accessor methods to the non-test code for each thing we need to
access.

This is done in anticipation of a test later in this series also
needing access to internals of the probe, via the fixture.