ct: Make dl_stm idempotent
The dl_stm_offsets object maintains three fields, last reconciled
offset, insync offset, and applied offset. Last two offsets are
used to implement concurrency control.

The insync_offset is a current position of the STM in the log.
The applied_offset is an offset of the record batch which was
applied to the STM last. When the stm manager invokes 'do_apply'
method of the 'dl_stm' the 'insync_offset' is updated to be equal
to the 'base_offset' of the record batch. If the record batch is
not a 'dl_stm_command' batch the method returns and the
'applied_offset' is not updated. But if the batch is a 'dl_stm_command'
batch the commands are applied to the in-memory state of the STM and
the 'applied_offset' is set to 'base_offset' of the record batch.
After that the 'applied_offset' is equal to 'insync_offset'.

Both 'insync_offset' and 'applied_offset' can't go back. They
can only increase with every update.

Before the commands could be applied the 'dl_stm' checks the
invariant. The invariant has two parts:
- the 'applied_offset' should be less than 'insync_offset'. This
  is because the code updates the 'insync_offset' first and then
  performs the check and in case of success updates the
  'applied_offset'. If the same batch is applied twice the
  'applied_offset' will be equal to 'insync_offset' and no changes
  will be made. This filters out possible duplicates (the STM sees
  one batch applied twice in a row).
- the 'insync_offset' should be equal to 'base_offset' of the command
  batch. This guarantees that the command can only be applied if it
  was never seen before. This handles situation when the log is
  replayed twice.

We need both invariants because 'dl_stm' uses 'base_offset' as
'version' and to track insync and applied offsets. The second check
alone will still allow the last batch to be applied twice.

The overlays and snapshots both contain 'version' fields which are
generated using 'base_offset' field of the applied 'dl_stm_command'
batch. This is useful because we can adjust the offsets when the Raft
snapshot is created. We can also filter the snapshots and overlays
based on log offsets. So essentially this gives us a snapshotting
mechanism for the 'dl_stm'.

The application of commands is updated to be deterministic. Since
the stm is idemptoent there is no need to handle duplicates in
'dl_stm_state'. Now it is safe to assume that if the command can't
be applied then it's incorrect and should be discarded.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>