`rptest`: use `RedpandaVersionTriple` in `java_compression_test.py`
The PR https://github.com/redpanda-data/redpanda/pull/25092 is being
backported to `v24.3` (see https://github.com/redpanda-data/redpanda/issues/25108).

Using a `RedpandaVersionLine` of `((24, 3))` will result in the latest available
version being used in the ducktape test in the future, and the upgrade test
expectation is that we should be using a version without the fix, which will
fail with snappy compression, before upgrading and succeeding.

Change the `pre_big_endian_snappy_fix_version` variable to use the value
`RedpandaVersionTriple((24, 3, 5))` to ensure we are using the correct version pre
`snappy` fix.