`rptest`: fix race in `compaction_recovery_test.py`
Attempting to delete segment indices for the list of segments
returned by `produce_until_segments()` is race-y, due to the fact
adjacent merge compaction could have kicked in while we were
waiting for the `redpanda` nodes to be stopped.

Requery it after the nodes have been restarted to ensure the segment
data is up-to-date.