k/server/tests: do not rely on chrono fmt behavior
Redpanda for a long time had it's own formatting for
std::chrono::millisecond via ostream insertion operator overload which
fmtlib used. Now fmt ships with its own implementation for formatting
chrono types (fmt/chrono.h) and end behavior depends on linking order.

In this test we either hard code the strings or call count() on duration
to get a numeric value the formatting of which I don't expect to change.