dt: test crash report telemetry
Add a ducktape test to verify that crash reports can be uploaded to a
test telemetry server successfully, and that the request bodies have the
expected json format.

The log allow lists in crash_loop_checks_test.py are exposed to allow
importing in the new tests. (Importing CrashLoopChecksTest would
reexecute the tests of CrashLoopChecksTest in the new test file as well.)

The mock telemetry server in metrics_reporter_test.py is reused in the
new test and adapted to specialize for heartbeat requests by default
(through its reports() method) while CrashReporterServer extends this
with the crash report-specific crash_reports() helper.