`rptest`: deflake `log_compaction_test.py`
This assert on `get_closed_segment_bytes()` is race-y in the parameterized
test which uses a `compact,delete` cleanup policy.

This is because it is possible for retention to kick in after waiting on
compaction, and fully remove all closed bytes in the log.

Change to only assert on this condition for `compact` topics to de-flake
the test.