iceberg: stop catalog client on shutdown
Previously the coordinator manager shutdown could take a while, notably
when a coordinator was in the middle of calling into the catalog. This
would make some tests with Iceberg topics timeout when shutting down.

This updates the catalog implementations to define a stop() method
(already defined for the REST catalog, but never called previously), and
makes the coordinator manager use it upon shutting down.