dl/coordinator: initial plumbing of cluster UUID
I have an upcoming change to use cluster UUIDs in the file committer.
This gets the initial plumbing out of the way.

Since the coordinator manager is constructed before the cluster UUID is
initialized, this required passing the storage::api. A subsequent commit
will use it to get the cluster UUID and use it in the file committer.