raft: pass in an abort source into replicate options
Added an option to pass an external abort source to
`raft::replicate_options`. This makes the Raft group lifecycle
management much easier as the `replicate` caller doesn't have to rely on
the Raft shutdown to abort waiting for replication to finish. That was
previously a major issue as the Raft group abort source had to be
triggered to interrupt waiting for replication result.

Now the caller can pass in an optional abort source reference to the
`replicate` request to interrupt waiting for replication.

Signed-off-by: Michał Maślanka <michal@redpanda.com>