model/compression: array of all compression modes
Add a constexpr array of all the compression modes applicable to batch
compression, so that we can loop over it or whatever we want. We can
only hold our breath that C++35 will offer something like this
built-in.

Testing:

Add some static asserts that the array of all compression types is
as expected (count member, front and back) and that all compression
types roundtrip through `fmt::format` and lexical_cast as expected.