rptest: memoize arch with a lock
In https://github.com/redpanda-data/redpanda/pull/25285 I introduced
multi threading for downloading Redpanda releases on all nodes in
parallel. This introduced a subtle race where the architecture was being
set to default (amd64) by on thread while fetching the actual
architecture which led to other threads believing that the work was
already done.

Do it under a lock now. One thread will fetch the actual value while
others will wait on the lock.

Also, don't set a default arch until we actually determine it. Without
this, if one thread gets an exception the others could continue
believing that the actual arch is amd64 while it is in fact arm64.