dl/translation/writer: simplify memory tracking
It appears that the approximate size per record is inaccurate compared
to what parquet is tracking and there is no guarantee that is smaller
than the parquet representation of bytes. This commit simplifies the memory
reservations to happen once the writer notifies the tracker of the
current usage at which point any deficit is reserved.

These changes means that we reserve the memory after appending a record
to parquet writer, so there is additionally one extra record that is
unaccounted for a brief window of period until the usage is reported.