`datalake`: set correct test settings in `compaction_test.py`
`cloud_storage_enable_remote_read` was meant to be set to `false`
in `CompactionTest`, not `CompactionGapsTest`, due to use of a
`KgoVerifierSeqConsumer` with `validate_latest_values=True`, which will
fail if the log consumed is not perfectly compacted.

For that reason, we need to ensure only the local log (which can guarantee
perfect compaction) is read, and not data from tiered storage (which may not
be compacted).