cluster/partition: Reword description for kafka_request_bytes_total.
This metric probably has the wrong name, the equivalent internal
metric is cluster_partition_request_bytes_total, the metric is
measuring disk activity, not kafka network traffic.

Reword the description to make it clearer.

The description is the same for both produce and consume because
only the first one is returned on the prometheus endpoint.

Signed-off-by: Ben Pope <ben@redpanda.com>