archival: New segment_collector mode
Adds new collector mode which can be used to upload new segments. This
makes 'segment_collector' suitable to replace some logic inside the
'archival_policy'. The long term goal here is to make 'segment_reupload'
the only code path which is used to find upload candidates. Then the
'ntp_archiver' should start using method of the segment collector which
returns a stream instead of the set of segments. When its done the
internals of the 'segment_collector' could be updated to use
'log_reader' instead of iterating through the segments.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>