k/group: added a conservative term check in group_manager
Added a check validating if partition term change when group state is
being validated. It is perfectly possible for the same node to become a
leader for the consumer group partition after it stepped down. Therefore
checking if a current replica is the leader is not enough. Term check
validates if current group state is fully up to date even if the replica
is still the leader.

Signed-off-by: Michał Maślanka <michal@redpanda.com>