c/controller: hold apply lock when starting shard balancer
Shard balancer is expected to start when there are no new updates to the
topic table, otherwise it might miss some notifications. Usually that
assumption is correct. The RPC services starts after the
`shard_balancer` is started so nothing new can be appended to the
controller log. The situation is different during the cluster bootstrap
where RPC services are started much earlier and the controller partition
may receive updates while the `shard_balancer` is starting. This
triggers an assertion in the `shard_balancer`. In order to allow the
balancer to operate in the state state we lock controller stm applies
for the time when balancer is started.

Signed-off-by: Michał Maślanka <michal@redpanda.com>