app: Add service shutdown control mechanism
Currently, the constructions of services is explicit in the code (we
have a set of functions which are invoking 'construct_service' for each
of them). But the shutdown process is not obvious. The application has a
queue. Every 'construct_service' call is adding a functor to the queue.
This functor is shutting down the service when invoked. The queue is
drained when the application stops. All functors are getting invoked in
order which is inverse of the order of construction.

Every service is responsible for its own logging and observability. Most
services are logging "Stopping .." on info level but not all. Because of
that sometimes it's tricky to debug shutdown problems (esp. if Redpanda
is logging on INFO level).

This commit fixes the problem by adding a genaralized mechanism that
controls how services are stopped. The 'construct_service' call is
setting up a shutdown callback which is logging the name of the service
before invoking the 'stop' method of the service (the method that can
potentially hang).

The callback is also setting up two watchodog timers before invoking
'stop'. The short watchdog is triggered after 30s. It logs message that
contains the name of the service that hods back the shutdown process on
the INFO level. The long watchdog is triggered after 3 minutes. It logs
similar message on WARN level.

The callback logs the name of the next service in the shutdown sequence
in the "Stopping ..." log messages. This simplifies the debugging
because you know what to grep to find the next step of the shutdown
procedure.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>