iceberg: check for JSON nulls when parsing optionals
We previously assumed that optional field would not be serialized rather
than having the be serialized as null. This is somewhat gray area in the
spec (e.g. previous impls would use -1 instead of null), but we have at
least seen one case (with Unity Catalog) where the metadata generated by
the catalog serializes 'current-snapshot-id' as null:

```
  ...
  "last-partition-id" : 1000,
  "default-sort-order-id" : 0,
  "sort-orders" : [ {
    "order-id" : 0,
    "fields" : [ ]
  } ],
  "properties" : {
    "write.object-storage.enabled" : "true",
    "write.metadata.compression-codec" : "gzip",
    "write.summary.partition-limit" : "100",
    "write.parquet.compression-codec" : "zstd"
  },
  "current-snapshot-id" : null,
  "refs" : { },
  "snapshots" : [ ],
  "statistics" : [ ],
  "partition-statistics" : [ ],
  "snapshot-log" : [ ],
  "metadata-log" : [ ]
}
```