`iceberg`: set `iceberg::field_required::yes` for `key_field` in `map_type`
Keys are required per the iceberg spec: https://iceberg.apache.org/spec/#nested-types.

Before this change, `redpanda` would fail to properly use schemas for tables
created with a `map` in the schema, and would incorrectly try to evolve
the schema due to the detected difference in `field_required`.

After this fix, schemas with nested `map` types work as expected.