rptest: Fix scaling_up_test.py
This commit fixes the ducktape test by adding a short sleep
between the deletion of the topic and recovery of the topic. When this
two operations are performed back to back it's possible for the recover
to pick up different version of the manifest on different replicas. One
of the replicas might be uploading the manifest concurrently while the
topic is deleted and created which will lead to divergence. This is not
a problem during the actual disaster recovery.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>