c/members_table: don't double-notify if node id was re-added
If the node id is present both in _nodes and _removed_nodes,
it means that the id was removed and later added (reverse order
is impossible, as the removal command removes the id from _nodes).
Therefore, when restoring from the controller snapshot, we only need
to notify such ids with the "added" state.

Fixes https://redpandadata.atlassian.net/browse/CORE-8625