fetch_bench: use right node_id
Fetch bench inexplicably takes 2 minutes to run, regardless of
iteration count or duration values. Root cause is that we try to move
the partition to "node 0, shard 0", under the assumption that this is
the node that is running the fetch bench. However, the right node is
actually 1 in this context (see redpanda_thread_fixture constructor).

This causes the partition movement to fail, but this is not detected
because we use an RPTEST_REQUIRE_EVENTUALLY_CORO which calls
BOOST_FAIL on timeout, which does nothing in an rpbench, so we continue
with no diagnostic after the movement failure.

This causes each test to hang for 30 seconds each, times 4 tests is
2 minutes.

Instead, query the right node directly from the node config. The test
now runs in ~1 second per test case.

Also, use cooperative_spin_wait_with_timeout directly instead of
the macro, so we will throw if it fails.