`storage`: add `disk_log_impl::find_adjacent_compaction_ranges()`
To extend adjacent segment compaction beyond a limit of two segments,
add a new function which will determine sub-arrays of valid segments
that can be concatenated together.

This sub-array is subject to much of the same limitations as before
with regards to adjacent segment compaction - raft terms must be identical,
total size pre-compaction should be less than `max_compacted_segment_size`,
and the offset space should not exceed the `uint32_t` limits per the
implementation of self compaction.