ct: Add extent_meta struct
The 'extent_meta' is a 'coordinate' of the data in the cloud storage. It
contain fields needed to reconstruct the object name (currently
only the id) and fields that can be used to locate the byte range inside
the object (offset and size).

It's only used internally by the data layer. In the current
implemenetation it replaces the lw_placeholder struct. The goal is to
decouple the data plane from the metadata layer. Currently the data
layer uses some data structures that describe the metadata
(dl_placeholder batch) directly.

The instances of this struct are supposed to be stored in memory.
They shouldn't be persisted by the metadata layer. They're only needed
to communicate with the data layer.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>