rpk: improve diagnostics steps error in debug bundle
rpk debug bundle works on a best-effort basis. It
always tries to return a bundle—even if some steps
(like logs or resource collection) fail.
Sometimes, is because the system is unhealthy
leading to expected errors. Seeing errors doesn't
mean the bundle is useless.

This change aims to make this clearer for the user.