cs/t: Force arbitrary disk stats in cloud_storage_e2e_test
Seems TestTimequeryAfterArchivalGC was racing against a disk status notification
for the cloud cache, occasionally leading to an exception in cache::trim.

Not entirely clear why this didn't flake in cmake land. This is fine as a hack
though, since the cache itself isn't under test. We're just trying to clear
it out to ensure we actually query block storage.