k/connection_ctx: wait for response future in foreground when stopping
When stopping Redpanda the server connection gate may be closed before
the `spawn_with_gate` is executed. In this case the future that is
captured in the lambda that is supposed to execute in the background is
not waited for. If the future failed it will trigger Seastar warning
about ignored failed future. To prevent this from happening added
waiting for result future to finish before returning from
`process_request` function.

Signed-off-by: Michał Maślanka <michal@redpanda.com>