`storage`: add `TestAdjacentCompactionSimulateCrashDuringFileRemoval` test
One of the concerns behind the new adjacent compaction routine is the
increased exposure to crashes while redundant data/`segment`s are being
permanently removed from disk.

While this was also a possibility with the old routine, the increased
exposure is due to the need to remove `N` `segment`s instead of just
1 with the previous restriction of 2 `segment`s per adjacent compact
operation.

Add a test that simulates a crash/failure to remove `segment`s from
disk after the concatenated `segment` is produced, leaving redundant
data on disk to see its effects on up-stream `segment` readers/consumers.