`test_utils`: set `--gtest_output` for `gtest_listener_test`
This test is expected to fail, but the `ASSERT_TRUE(false)` call will
result in `gtest` & `bazel` writing a failure to `test.xml`.

To avoid this, redirect the `xml` output to `/dev/null` so the test
appears as a pass when the `test.xml` file is read by our CI scripts.