rpk: add logger to SR client
We are bumping franz-go to pull the recent changes
that introduce the logFn to the SR package.

This allow us to have a logger in our schema
registry client.

The wrapKgoLogger function can be removed once
franz-go 1.20 is released as this will be a
helper function in the kgo package.

Newer franz-go requires 1.23.8; we are bumping
Go to 1.24.3 now.

Also, franz-go introduced lag calculation for
non-empty groups which made us change the
offset_deletion test and make it more reliable
with the new rpk behavior.

Fixes UX-24