`config`: add `log_compaction_pause_use_sliding_window`
We have seen in incidents that it can be useful to disable sliding
window compaction for a period of time, in order to allow adjacent
segment compaction to kick in instead.

Unfortunately, `log_compaction_use_sliding_window` is marked as
`needs_restart::yes`, due to the fact memory is only allocated at start-up
if it is enabled.

To deal with the common case in which we start `redpanda` with
`log_compaction_use_sliding_window=true`, and wish to disable sliding window
compaction for a period of time without forcing a cluster restart (usually by
cloud operators), add a new cluster config `log_compaction_pause_use_sliding_window`,
which doesn't require a restart and can be used to temporarily disable sliding
window compaction.