cloud: parameterize credentials refresh by AWS service
Credentials for signing requests to Glue will also need to be
periodically refreshed. This change updates the credentials refresh
machinery so later it can be used with Glue.

It's a bit ad hoc, but all of this code is bit ad hoc because of how
it's trying to handle multiple cloud providers and mechanisms with a
uniform abstraction.