utils: Update retry_chain_node
Add 'retry_chain_context' class. The context stores the reference to the
abort source so it could be used to abort the computation.

Also, the retry_chain_context can accumulate trace logs using limited
amount of memory. The traces could be retrieved. The context has a
'reset' method that removes all accumulated traces.

The 'retry_chain_logger' can interact with the context to save the
traces. The idea is that we can dump all trace logs in case of stall or
unexpected failure.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>