`rptest`: deflake `datalake/compaction_test.py`
It seems that the number of dirty bytes in the log was not a air-tight
indicator of compaction being finished on its own. This is due to the restart
of all of the nodes in the cluster being issued right before a call to
`redpanda.metric_sum()`. This can be race-y and return `0` immediately after
a restart, before the log has a chance to book-keep the correct value for
dirty bytes given the existing segments on disk.

Re-use a proven metric, that being the number of complete sliding window
compaction rounds, and sample it over multiple periods of
`log_compaction_interval_ms` to be sure that compaction has settled
before attempting to verify the compacted log with `KgoSeqConsumer`.