schema_registry: acl request json ser/de
Define the acl types used for parsing and serializing ACL CRUD requests
and responses. Define the json serde methods on it and add a conversion
method from acl_bindings.

The `acl` holder state is reused as the representation of GET response
values, POST request values and DELETE request values as well as DELETE
response values to reduce duplication.