rptest: deterministic cloud storage usage test
With a constant stream of data arriving on the topic it is very hard to
ensure that admin endpoint of Redpanda and our python test will see a
consistent view of the manifest across multiple partitions. It usually
works but the test is flaky since the first day.

Rewrite it such that we produce some data, then we wait for admin report
to converge to the report as seen by the python script.