rptest: don't query unstarted nodes for metrics
Recently we changed the rptest wrapper to query each node during
stop_node for metrics. However, it turns out that stop_node is called
by DT itself as part of clean_node, which always results in a failed
metric query and a bunch of spam to the warning log (which shows up
on the console in DT).

Check if the node is started before querying it to avoid this.