security/acl: envelope resource_pattern_filter
resource_pattern_filter was previously serde serialized using the
low-level read_nested and write methods. This meant that it was
serialized without an envelope. This makes it difficult to extend it
with new fields in a backwards compatible way.

To allow adding new fields into it, we introduce a version v1 of it,
that uses serde_fields to serialize it with an envelope header as
expected. However, v1 is not compatible with v0.

We rely on the serde of the containing type to double write the field
and allow switching over to using v1.
See the `acl_binding_filter::serde_write` for implementation details of
this switchover.