c/data_migrations: fill topic locations for outbound migrations
Topic locations are filled from topic table information when the create
migration command is applied to the migration table. This means that
topic location info is duplicated, but the advantage of saving it in the
migration table as well is that it is still available after local topics
are deleted.