`storage`: set cached disk usage in `do_swap_data_file_handles()`
The major motivation for these changes is the number of calls being made
to `fstat()` for each `segment` during a housekeeping round. These syscalls
must go through the seastar `syscall_work_queue`, and take up much more
CPU utilization than they should when studying a flamegraph.

The common thread in all of the flamegraph stack traces is
`segment::persistent_size()`, which issues `fstat()` commands to acquire
the size of the `segment`, `segment_index` and `compacted_index` if they
are not currently cached.

We issue calls to this function from within `log_manager::gc_loop()`, which
queries a `log` for its `disk_usage()`, which must make calls to EACH
`segment`'s `persistent_size()`.

It is clear to see that compaction constantly invalidating these values
and then not re-assigning them could result in a lot of time spent in syscalls
to disk. Thankfully, we have enough state in memory to reassign (at the very least,
a rough estimation) the size instead of just clearing the cache.

After plumbing the returned value of `new_cmp_idx_size` throughout
the compaction utils layer, we can now call `s->set_cached_disk_usage()`
after clearing it within `do_swap_data_file_handles()`. The size of the `segment`
is set according to the `segment_reader::file_size()`, which is guaranteed to
be correct, as a call to `segment_reader::load_size()` is issued right above it
in the function. The `segment_index`'s size will be assigned instead when a call
to `segment_index::flush()` is made.

This should all have the intended effect of cutting down on system calls in the
housekeeping layer.