rptest/datalake: log metrics query results in flaky test
DatalakeCustomPartitioningTest test_basic is sometimes flaky because it
sees too few translations and files created in the metrics query, even
though the query occurs long after translations are finished and uploads
are done, and there are no signs of metrics errors or any other cause of
the discrepancy. As a first step towards debugging this issue, log the
results of the metrics query for each node at the debug level. This
should narrow down the problem to a particular node, which will make
correlation with the logs easier.