dt/keycloak: increase start timeout to 300s
The audit logging tests have been flaky because of timeouts on starting
up the keycloak service in CI in ducktape.

Looking through recent passing and failing tests, it is clear that the
startup time of keycloak varies significantly between 20s and 120s+.
Unfortunately keycloak is an external service and the logs are not
helpful to debug where exactly this wide variant is stemming from, and
therefore it is difficult to see if and how it may be reduced.

This commit gives keycloak more time to start to ensure that it can
start up even at the upper end of its range, to reduce flakiness for the
tests that use keycloak.