serde/json: parse any JSON value at top-level
The extended JSON spec allows parsing any JSON value at top-level. This
changes extends the parser to do this. It's almost a trivial change, but
parsing top-level numbers is a bit tricky, because the numeric parser
can be suspended after any byte, so it cannot tell us if the JSON string
"42" is complete or not. To work around it, we do a little trick and, if
at the EOF and top-level, feed the number parser a phantom space, which
will trick it into recognizing if it's seen a complete literal or not.